requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0

# RAG System
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
pypdf>=3.17.0

# Document processing
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
unstructured>=0.11.0

# Phoenix Evals
arize-phoenix>=4.0.0
arize-phoenix-evals>=0.10.0

# Utilities
tqdm>=4.66.0
pandas>=2.0.0
numpy>=1.24.0
